src/requirements.txt
pillow
numpy
flask
requests
img2vec_pytorch
pytest